minor edit MainLayout. create LayoutHeader component
